move mask inside module, demo and mask cmds, display logging on cmdline
