+++ scripts/pybayesx/pybayesx/scripts.py
@@ -1,14 +1,19 @@
-from argparse import ArgumentParser, Namespace
+import logging
+from argparse import ArgumentParser, FileType, Namespace
 from pathlib import Path
+from sys import stderr, stdout
 
 from getdist import loadMCSamples
 
+from .analysis import demo
+from .mask import mask
 from .plot import plot
+
+log = logging.getLogger(__name__)
 
 
 def main():
     parser = ArgumentParser(description="Basic interaction with BayesX")
-
     subparsers = parser.add_subparsers(required=True, help="sub-command")
 
     # Plot subparser
@@ -36,6 +41,60 @@ def main():
     )
     plot_parser.set_defaults(func=interactive_plot)
 
+    # Mask subparser
+    mask_parser = subparsers.add_parser(
+        "mask",
+        description="Generate mask for given files of ellipses generated by CIAO",
+    )
+    bounds = mask_parser.add_argument_group("bounds", "Coordinates of data bounds.")
+    bounds.add_argument("xMin", type=float, help="Minimum x coordinate in data.")
+    bounds.add_argument("yMin", type=float, help="Minimum y coordinate in data.")
+    bounds.add_argument("xMax", type=float, help="Maximum x coordinate in data.")
+    bounds.add_argument("yMax", type=float, help="Maximum x coordinate in data.")
+    mask_parser.add_argument(
+        "maskFiles", nargs="+", help="List of paths to files of ellipses to mask"
+    )
+    # Optional
+    mask_parser.add_argument(
+        "-d",
+        "--overdraw",
+        help="Positive values increase effective radius of ellipse, negative decrease.",
+        type=float,
+        default=0,
+    )
+    mask_parser.add_argument(
+        "-o",
+        "--outputFile",
+        help=(
+            "Path to output file. File extensions '.txt' and '.txt.gz' will export to "
+            "a text file, with compression in the latter case. This is slowest and "
+            "large. Best for human readable output. '.npy' will export to an "
+            "uncompressed numpy .npy file. This is the fastest option but large. "
+            "'.npz' (default) will export to a compressed numpy .npz file. This is "
+            "slower than .npy but much smaller. '.matxt' will export entire mask matrix"
+            " to text file.  This is intended for debugging purposes and is not "
+            "formatted for the binning script.",
+        ),  # type: ignore
+        type=FileType("wb"),
+        default="mask.npz",
+    )
+    mask_parser.set_defaults(func=_mask)
+
+    # Demo parser
+    demo_parser = subparsers.add_parser("demo", description="Run demo analysis.")
+    demo_parser.set_defaults(func=_demo)
+
+    # Write logging to shell
+    rlog = logging.getLogger()
+    rlog.setLevel(logging.INFO)  # log everything
+    infoHandler = logging.StreamHandler(stream=stdout)
+    infoHandler.setLevel(logging.INFO)  # write info to stdout
+    rlog.addHandler(infoHandler)
+    errorHandler = logging.StreamHandler(stream=stderr)
+    errorHandler.setLevel(logging.WARNING)  # warnings or errors go to stderr
+    rlog.addHandler(errorHandler)
+
+    # Call appropriate function for subcommand
     args = parser.parse_args()
     args.func(args)
 
@@ -85,3 +144,11 @@ def interactive_plot(args: Namespace):
         plot_file=args.output,
         chain_labels=args.label,
     )
+
+
+def _mask(args):
+    mask(args)
+
+
+def _demo(args):
+    demo()
